test-gml-code/pyproject.toml
[project]
name = "test-gml-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "zai-sdk==0.0.4",
    "openai",
]